+++ data/data_8107224.s
@@ -1,16 +1,13 @@
 .section .rodata
 
 
-.string "pksdir0\0" @Groudon
-.string "pksdir0\0" @Ray
+
+.string "pksdir0\0" @Mewtwo
 .string "pksdir0\0" @Entei
 .string "pksdir0\0" @Raikou
 .string "pksdir0\0" @Suicune
 .string "pksdir0\0" @HoOh
 .string "pksdir0\0" @Latios
-.string "pksdir0\0"
-.string "pksdir0\0"
-.string "pksdir0\0"
-.string "pksdir0\0" @ 
+.string "pksdir0\0" @ regis
 
 .align 2,0
