+++ src/dungeon_cutscene_rayquaza.c
@@ -1,0 +1,255 @@
+#include "global.h"
+#include "globaldata.h"
+#include "code_800DAC0.h"
+#include "code_800E9A8.h"
+#include "code_803E46C.h"
+#include "code_803E668.h"
+#include "code_8041AD0.h"
+#include "code_804267C.h"
+#include "code_80861A8.h"
+#include "code_806CD90.h"
+#include "code_80869E4.h"
+#include "constants/bg_music.h"
+#include "constants/direction.h"
+#include "constants/friend_area.h"
+#include "constants/weather.h"
+#include "dungeon_items.h"
+#include "dungeon_leader.h"
+#include "dungeon_map_access.h"
+#include "dungeon_message.h"
+#include "dungeon_music.h"
+#include "dungeon_misc.h"
+#include "dungeon_logic.h"
+#include "dungeon_random.h"
+#include "dungeon_util_1.h"
+#include "dungeon_util.h"
+#include "exclusive_pokemon.h"
+#include "friend_area.h"
+#include "items.h"
+#include "pokemon_3.h"
+#include "pokemon.h"
+#include "position_util.h"
+#include "trap.h"
+#include "math.h"
+#include "dungeon_config.h"
+#include "dungeon_boss_dialogue.h"
+
+extern void sub_8042B0C(Entity *);
+extern u8 sub_806FD18(Entity *);
+extern void sub_806FDF4(Entity *, Entity *, Entity **);
+extern s32 GetCameraXPos(void);
+extern s32 GetCameraYPos(void);
+extern void sub_803F878(u32, s32);
+extern void BgColorCallNullsub4(void);
+extern void sub_8085EB0(void);
+extern void sub_8049884(void);
+extern void sub_8049B8C(void);
+extern void ShowWholeRevealedDungeonMap(void);
+extern void sub_8086A54(Entity *);
+extern void ResetMonEntityData(EntityInfo *, u32);
+extern void sub_8041888(u32);
+extern u32 sub_80861F8(u32, Entity *, u32);
+extern u8 sub_80860A8(u32);
+extern void sub_8052D44(s16 *, Entity *, Entity *);
+extern void sub_80464C8(Entity *, DungeonPos *, Item *);
+extern void SetDungeonBGColorRGB(u32, u32, u32, u32, u32);
+extern u32 sub_8085EC8(u32, u32, u32, DungeonPos *, u32);
+extern void sub_807EAA0(u32, u32);
+extern void sub_8072008(Entity *, Entity *, s16, u32, u32);
+extern void sub_8085374(void);
+extern u32 GetRandomFloorItem(u32);
+extern bool8 sub_8085B80(struct_8085B80 *);
+extern void sub_8086A3C(Entity *pokemon);
+extern void SetupBossFightHP(Entity *pokemon, s32 newHP, u16 songIndex);
+extern u8 sub_8086AE4(s16 _index);
+
+static void RayquazaDropInEffect(Entity *rayquazaEntity);
+static void RayquazaScreenFlash(void);
+
+void sub_80892C8(void)
+{
+  Entity * leaderEntity;
+  Entity * RayquazaEntity;
+
+  leaderEntity = xxx_call_GetLeader();
+  RayquazaEntity = GetEntityFromMonsterBehavior(BEHAVIOR_RAYQUAZA);
+  DungeonStartNewBGM(MUS_RAYQUAZAS_DOMAIN);
+  sub_8085374();
+  sub_80854D4();
+  sub_8085930(DIRECTION_NORTH);
+  sub_80855E4(sub_8086A3C);
+  SetFacingDirection(RayquazaEntity, DIRECTION_SOUTH);
+  sub_8086A3C(RayquazaEntity);
+  sub_8085860(leaderEntity->pos.x,leaderEntity->pos.y - 3);
+  CopyMonsterNameToBuffer(gFormatBuffer_Monsters[2],MONSTER_RAYQUAZA);
+}
+
+void sub_8089328(void)
+{
+  Entity * leaderEntity;
+  Entity * rayquazaEntity;
+
+  leaderEntity = xxx_call_GetLeader();
+  rayquazaEntity = GetEntityFromMonsterBehavior(BEHAVIOR_RAYQUAZA);
+  DungeonStartNewBGM(MUS_RAYQUAZAS_DOMAIN);
+  sub_80854D4();
+  sub_8085930(DIRECTION_NORTH);
+  sub_80855E4(sub_8086A3C);
+  if (sub_8086AE4(MONSTER_RAYQUAZA)) {
+    sub_8068FE0(rayquazaEntity,0x21c,0);
+  }
+  else {
+    sub_8072008(rayquazaEntity,rayquazaEntity,gRayquazaConfigLevel,0,0);
+    sub_8086A3C(rayquazaEntity);
+  }
+  sub_8085860(leaderEntity->pos.x,leaderEntity->pos.y - 3);
+  CopyMonsterNameToBuffer(gFormatBuffer_Monsters[2],MONSTER_RAYQUAZA);
+}
+
+void sub_80893B4(u8 param_1, u8 param_2)
+{
+  if ((param_2 == 0x17 || param_2 == 0x18 || param_2 == 0x19) && (param_1 == 0xF)) {
+    sub_8097FA8(0xF);
+    gDungeon->unk2 = 1;
+  }
+}
+
+void RayquazaPreFightDialogue(void)
+{
+  Entity *leaderEntity;
+  Entity *partnerEntity;
+  Entity *rayquazaEntity;
+
+  leaderEntity = xxx_call_GetLeader();
+  partnerEntity = GetPartnerEntity();
+  rayquazaEntity = GetEntityFromMonsterBehavior(BEHAVIOR_RAYQUAZA);
+  sub_8086448();
+  DisplayDungeonDialogue(&gRayquazaPreFightDialogue_1);
+  RayquazaScreenFlash();
+  DisplayDungeonDialogue(&gRayquazaPreFightDialogue_2);
+  sub_80856E0(partnerEntity, DIRECTION_NORTH);
+  SpriteLookAroundEffect(partnerEntity);
+  sub_803E708(10,0x46);
+  sub_80869E4(partnerEntity,4,1,DIRECTION_EAST);
+  sub_80869E4(leaderEntity,4,2,DIRECTION_WEST);
+  DisplayDungeonDialogue(gRayquazaPreFightDialogue_3);
+  sub_80869E4(partnerEntity,4,2,DIRECTION_NORTH);
+  sub_80869E4(leaderEntity,4,1,DIRECTION_NORTH);
+  DisplayDungeonDialogue(&gRayquazaPreFightDialogue_4);
+  sub_803E708(10,0x46);
+  DisplayDungeonDialogue(&gRayquazaPreFightDialogue_5);
+  sub_803E708(10,0x46);
+  DisplayDungeonDialogue(&gRayquazaPreFightDialogue_6);
+  RayquazaScreenFlash();
+  PlaySoundEffect(0x1da); // Rayquaza Cry
+  DisplayDungeonDialogue(&gRayquazaPreFightDialogue_7);
+  RayquazaDropInEffect(rayquazaEntity);
+  DisplayDungeonDialogue(&gRayquazaPreFightDialogue_8);
+  RayquazaScreenFlash();
+  DisplayDungeonDialogue(&gRayquazaPreFightDialogue_9);
+  RayquazaScreenFlash();
+  DisplayDungeonDialogue(&gRayquazaPreFightDialogue_10);
+  RayquazaScreenFlash();
+  DisplayDungeonDialogue(&gRayquazaPreFightDialogue_11);
+  SetupBossFightHP(rayquazaEntity,600,MUS_BATTLE_WITH_RAYQUAZA);
+  ShiftCameraToPosition(&leaderEntity->pixelPos,0x10);
+}
+
+void RayquazaReFightDialogue(void)
+{
+  Entity *leaderEntity;
+  Entity *rayquazaEntity;
+
+  leaderEntity = xxx_call_GetLeader();
+  rayquazaEntity = GetEntityFromMonsterBehavior(BEHAVIOR_RAYQUAZA);
+  sub_8086448();
+  RayquazaScreenFlash();
+  PlaySoundEffect(0x1da); // Rayquaza Cry
+  DisplayDungeonDialogue(&gRayquazaPreFightDialogue_7);
+  sub_803E708(10,0x46);
+  RayquazaDropInEffect(rayquazaEntity);
+  DisplayDungeonDialogue(&gRayquazaReFightDialogue_1);
+  RayquazaScreenFlash();
+  DisplayDungeonDialogue(&gRayquazaReFightDialogue_2);
+  RayquazaScreenFlash();
+  DisplayDungeonDialogue(&gRayquazaReFightDialogue_3);
+  RayquazaScreenFlash();
+  DisplayDungeonDialogue(&gRayquazaReFightDialogue_4);
+  SetupBossFightHP(rayquazaEntity,600,MUS_BATTLE_WITH_RAYQUAZA);
+  ShiftCameraToPosition(&leaderEntity->pixelPos,0x10);
+}
+
+void RayquazaPostStoryPreFightDialogue(void)
+{
+  Entity *leaderEntity;
+  Entity *rayquazaEntity;
+
+  leaderEntity = xxx_call_GetLeader();
+  rayquazaEntity = GetEntityFromMonsterBehavior(BEHAVIOR_RAYQUAZA);
+  sub_8086448();
+  if (sub_8086AE4(MONSTER_RAYQUAZA)) {
+    sub_80866C4(&gRayquazaPostStoryPreFightDialogue_6);
+  }
+  else {
+    PlaySoundEffect(0x1da); // Rayquaza Cry
+    DisplayDungeonDialogue(&gRayquazaPostStoryPreFightDialogue_1);
+    RayquazaDropInEffect(rayquazaEntity);
+    DisplayDungeonDialogue(&gRayquazaPostStoryPreFightDialogue_2);
+    RayquazaScreenFlash();
+    DisplayDungeonDialogue(&gRayquazaPostStoryPreFightDialogue_3);
+    RayquazaScreenFlash();
+    DisplayDungeonDialogue(&gRayquazaPostStoryPreFightDialogue_4);
+    RayquazaScreenFlash();
+    DisplayDungeonDialogue(&gRayquazaPostStoryPreFightDialogue_5);
+    SetupBossFightHP(rayquazaEntity,600,MUS_BOSS_BATTLE);
+    ShiftCameraToPosition(&leaderEntity->pixelPos,0x10);
+  }
+}
+
+static void RayquazaDropInEffect(Entity *rayquazaEntity)
+{
+  s24_8 iVar1;
+  s24_8 iVar2;
+
+  GetEntInfo(rayquazaEntity)->unk15E = 0;
+  iVar2 = IntToF248_2(200);
+  iVar1 = IntToF248_2(6);
+  PlaySoundEffect(0x1f8);
+  while( 1 ) {
+    iVar2.raw = iVar2.raw - iVar1.raw; // must be .raw
+    iVar1 = F248_Sub(iVar1, FloatToF248_2(3./32.));
+    if (F248LessThanFloat(iVar1, 0.08)) {
+      iVar1 = FloatToF248(0.08);
+    }
+    if (F248LessThanInt(iVar2, 0)) break;
+    GetEntInfo(rayquazaEntity)->unk174 = iVar2;
+    DungeonRunFrameActions(0x46);
+  }
+  GetEntInfo(rayquazaEntity)->unk174 = IntToF248_2(0);
+}
+
+static void RayquazaScreenFlash(void)
+{
+  s32 iVar1;
+
+  PlaySoundEffect(0x1f6);
+  for(iVar1 = 250; iVar1 > 149; iVar1 -= 10)
+  {
+    SetDungeonBGColorRGB(iVar1 / 2, iVar1, iVar1 / 2, 1, 1);
+    DungeonRunFrameActions(70);
+  }
+  sub_803E708(10,70);
+  for(iVar1 = 250; iVar1 > 199; iVar1 -= 10)
+  {
+    SetDungeonBGColorRGB(iVar1 / 2, iVar1, 0, 1, 1);
+    DungeonRunFrameActions(70);
+  }
+  sub_803E708(10,70);
+  for(iVar1 = 250; iVar1 >= 0; iVar1 -= 10)
+  {
+    SetDungeonBGColorRGB(iVar1 / 2, iVar1, iVar1 / 2, 1, 1);
+    DungeonRunFrameActions(70);
+  }
+  sub_803E708(10,70);
+  sub_8085EB0();
+}
